requirements.txt
openai
requests
python-dotenv
pydantic-ai
aiohttp
motor